display: move `Image` to `components/image`
